Arrange test locally
