+++ tests/integration/any_charm.py
@@ -1,35 +1,15 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+# pylint: disable=import-error,consider-using-with,no-member,too-few-public-methods
+
 """This code should be loaded into any-charm which is used for integration tests."""
-
-
-def install(package):
-    """Install."""
-    import importlib
-
-    try:
-        importlib.import_module(package)
-    except ImportError:
-        import pip
-
-        pip.main(["install", package])
-    finally:
-        globals()[package] = importlib.import_module(package)
-
-
-install("pydantic")
 
 import logging
 import uuid
 
 from any_charm_base import AnyCharmBase
-from dns_record import (
-    DNSRecordRequestProcessed,
-    DNSRecordRequirerData,
-    DNSRecordRequires,
-    RequirerEntry,
-)
+from dns_record import DNSRecordRequirerData, DNSRecordRequires, RequirerEntry
 
 logger = logging.getLogger(__name__)
 
@@ -47,15 +27,6 @@ class AnyCharm(AnyCharmBase):
         super().__init__(*args, **kwargs)
         self.dns_record = DNSRecordRequires(self)
         self.framework.observe(
-            self.dns_record.on.dns_record_request_processed, self._on_dns_record_request_processed
-        )
-        # self.framework.observe(
-        #     self.dns_record.on.dns_record_relation_changed, self._on_dns_record_relation_changed
-        # )
-        # self.framework.observe(
-        #     self.on.dns_record_relation_changed, self._on_dns_record_relation_changed2
-        # )
-        self.framework.observe(
             self.on.dns_record_relation_joined, self._on_dns_record_relation_joined
         )
 
@@ -71,21 +42,9 @@ class AnyCharm(AnyCharmBase):
         )
         dns_record_requirer_data = DNSRecordRequirerData(
             dns_entries=[dns_entry],
-            service_account="123213123123123123123",
+            service_account="fakeserviceaccount",
         )
         return dns_record_requirer_data
 
     def _on_dns_record_relation_joined(self, event) -> None:
-        logger.debug("JOINED")
         self.dns_record.update_relation_data(event.relation, self._test_record_data())
-
-    def _on_dns_record_relation_changed2(self, event) -> None:
-        logger.debug("CHANGED2")
-        self.dns_record.update_relation_data(event.relation, self._test_record_data())
-
-    def _on_dns_record_relation_changed(self, event) -> None:
-        logger.debug("CHANGED")
-        self.dns_record.update_relation_data(event.relation, self._test_record_data())
-
-    def _on_dns_record_request_processed(self, event: DNSRecordRequestProcessed) -> None:
-        logger.debug("PROCESSED")
